add text scale slider to adjust text size
