+++ lib/screens/settings/components/settings_title.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class SettingsTitle extends StatelessWidget {
+  final String title;
+
+  const SettingsTitle({super.key, required this.title});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 16),
+      child: Text(
+        title,
+        style: Theme.of(context).textTheme.labelMedium?.copyWith(
+              fontWeight: FontWeight.bold,
+              color: Theme.of(context).colorScheme.onSurface,
+            ),
+      ),
+    );
+  }
+}
